+++ limbook_api/models/__init__.py
@@ -1,3 +1,7 @@
+# This should be at the top
+from limbook_api.models.setup import *
+
+# Other models
 from limbook_api.models.image import *
 from limbook_api.models.activity import *
 from limbook_api.models.comment import *
