+++ seeds/demo.py
@@ -2,7 +2,7 @@ from faker import Faker as Generator
 from flask_seeder import Seeder
 
 from limbook_api.models import create_post
-from limbook_api.setup_db import db_drop_and_create_all
+from limbook_api.models import db_drop_and_create_all
 
 generator = Generator()
